requirements.txt
PySide6>=6.8.0
openpyxl>=3.1.0
python-docx>=1.1.0
pyinstaller>=6.0.0
